// src/lib/api/generated/models/subscriptionDetailResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */
import type { ResourceId } from './resourceId';
import type { SubscriptionDetailResponseBuyer } from './subscriptionDetailResponseBuyer';
import type { SubscriptionDetailResponseProduct } from './subscriptionDetailResponseProduct';
import type { SubscriptionDetailResponseSeller } from './subscriptionDetailResponseSeller';
import type { SubscriptionStatus } from './subscriptionStatus';
import type { UserId } from './userId';

export interface SubscriptionDetailResponse {
  id: ResourceId;
  buyerId: UserId;
  productId: ResourceId;
  sellerId: UserId;
  quantityOz: string;
  status: SubscriptionStatus;
  /** @nullable */
  cancelReason: string | null;
  fulfillmentType: string;
  /** @nullable */
  nextDeliveryDate: string | null;
  /** @nullable */
  createdAt: string | null;
  /** @nullable */
  updatedAt: string | null;
  product: SubscriptionDetailResponseProduct;
  /** @nullable */
  buyer: SubscriptionDetailResponseBuyer;
  /** @nullable */
  seller: SubscriptionDetailResponseSeller;
}